Removed unneeded static_cast statement due to engine modifications

+++ src/LabelWidget.cc
@@ -111,7 +111,7 @@ namespace sdl {
           break;
       }
 
-      utils::Boxf dstRect(center, static_cast<float>(sizeText.w()), static_cast<float>(sizeText.h()));
+      utils::Boxf dstRect(center, sizeText.w(), sizeText.h());
 
       // Compute the intersection between the input `area` and this `dstRect`
       // area. If both overlaps it means that part of the text is visible.
